requirements.txt
pyyaml==6.0
flask==2.3.0
requests==2.31.0
pydantic==2.0.0
psycopg2-binary==2.9.0
redis==5.0.0
scikit-learn==1.3.0
numpy==1.24.0
pytest==7.4.0
